+++ account/views/otp.py
@@ -22,7 +22,7 @@ from django.contrib import messages
 from django.core.exceptions import ValidationError, SuspiciousOperation
 from django.utils.translation import ugettext as _
 from django_otp.decorators import otp_required
-from sudo.decorators import sudo_required
+from elevate.decorators import elevate_required
 from two_factor import forms as two_forms
 from two_factor.views import core, profile
 
@@ -81,9 +81,9 @@ class TwoFactorSetupView(core.SetupView):
             raise SuspiciousOperation("ManagementForm data is missing or has been tampered.")
 
 
-backup_view = sudo_required(core.BackupTokensView.as_view(template_name="account/twofactor-backup.html", success_url="user-twofactor-backup"))
-disable_view = sudo_required(otp_required(profile.DisableView.as_view(template_name="account/twofactor-disable.html", success_url="user-security")))
+backup_view = elevate_required(core.BackupTokensView.as_view(template_name="account/twofactor-backup.html", success_url="user-twofactor-backup"))
+disable_view = elevate_required(otp_required(profile.DisableView.as_view(template_name="account/twofactor-disable.html", success_url="user-security")))
 login = anonymous_required(LoginView.as_view())
-setup_view = sudo_required(TwoFactorSetupView.as_view())
-qrcode_view = sudo_required(core.QRGeneratorView.as_view())
+setup_view = elevate_required(TwoFactorSetupView.as_view())
+qrcode_view = elevate_required(core.QRGeneratorView.as_view())
 twofactor_view = profile.ProfileView.as_view(template_name="account/security.html")