+++ account/views/settings.py
@@ -21,7 +21,7 @@ from django.core.urlresolvers import reverse_lazy
 from django.views import generic
 
 from braces.views import LoginRequiredMixin
-from sudo.mixins import SudoMixin
+from elevate.mixins import ElevateMixin
 
 from account import forms
 
@@ -45,7 +45,7 @@ class GeneralSettingsView(LoginRequiredMixin, generic.FormView):
         return super(GeneralSettingsView, self).form_valid(form=form, *args, **kwargs)
 
 
-class UsernameChangeView(LoginRequiredMixin, SudoMixin, generic.FormView):
+class UsernameChangeView(LoginRequiredMixin, ElevateMixin, generic.FormView):
     """Allow users to change their username"""
     form_class = forms.UsernameChangeForm
     success_url = reverse_lazy("user-settings")