+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Tests for OTP views (#283)
 * Buttons should give some visual feedback that they're doing something (#279)
 * Initial JS tests (#303)
+* Move from django-sudo to django-elevate (#282)
 * Reduce static asset sizes
   * Use UglifyJS to mangle our JS reducing each of our JS bundles by about 50KB
   * Only import the parts of Bootstrap that we actually use, reducing our CSS bundle by about 40KB